requirements.txt
numpy>=1.16
opencv-python>=3.3,<4.0
tensorflow-gpu>=1.12.0
Keras>=2.2.3
